third_party/WebKit/LayoutTests/http/tests/security/naked-global-object.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
    window.jsTestIsAsync = true;

    window.callback = null;

    function doCallback()
    {
        shouldBeUndefined(window.callback());
        finishJSTest();
    }
        
    function registerCallback(cb)
    {
        window.callback = cb;
        setTimeout(function() {
            var iframe = document.querySelector('iframe');
            iframe.onload = doCallback;
            iframe.contentDocument.location.href = "about:blank";
        }, 0);
    }
</script>
<iframe src="resources/iframe-for-naked-global-object.html"></iframe>
